klog: fix klogf va_arg for i16 — %x/%u read unsigned int, %lx/%lu for 32-bit
On i16, unsigned int is 16-bit.  klogf's %x/%u used va_arg(ap, uint32_t)
which reads 4 bytes even when callers push only 2 — corrupting all
subsequent arguments and producing garbled boot log output.

Fix: %x/%u now read unsigned int (platform-native width), %lx/%lu
read unsigned long (32-bit on all platforms).  This matches standard
printf conventions and is a no-op on 32-bit targets where
unsigned int == uint32_t.

Updated ~130 format specifiers across 13 files to use %lx/%lu for
explicitly 32-bit values (addresses, register dumps, sizes).

+++ src/kernel/klog.c
@@ -113,10 +113,10 @@ void klogf(const char *fmt, ...) {
           klog_puts_raw(va_arg(ap, const char *));
           break;
         case 'u':
-          klog_print_dec(va_arg(ap, uint32_t));
+          klog_print_dec((uint32_t)va_arg(ap, unsigned int));
           break;
         case 'x':
-          klog_print_hex32(va_arg(ap, uint32_t));
+          klog_print_hex32((uint32_t)va_arg(ap, unsigned int));
           break;
         case 'l':
           fmt++;
